requirements.txt
deepchem
fcd==1.2
guacamol
ipykernel
jupyterlab
matplotlib
numpy<2.0
pandas==1.5.3
rdkit==2023.9.2
scipy==1.10.1
seaborn
tqdm
transformers
wandb
tensorflow==2.9.1
molecule-generation
unimol_tools
huggingface_hub
einops
torchvision